pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "esm-catalog"
version = "0.1.0"
description = "STAC catalog generator for ESM Tools experiment outputs"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.0",
    "loguru>=0.7",
    "pydantic>=2.0",
    "pystac>=1.8",
    "xarray",
    "netcdf4",
    "pyyaml",
    "numpy",
    "snakemake>=8.0",
]

[project.optional-dependencies]
serve = [
    "fastapi",
    "uvicorn",
    "stac-fastapi.api",
    "stac-fastapi.types",
    "stac-fastapi.extensions",
    "attrs",
]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
esm_catalog = "esm_catalog.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/esm_catalog"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build.targets.sdist]
include = ["src/esm_catalog/**/*"]

[tool.hatch.build]
include = ["src/esm_catalog/**/*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
